# dist/swagger/swagger-output.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "판다마켓 API 문서(by 조형민)",
        "description": "코드잇 스프린트 FS4 스프린트 미션 10에서 진행한 swagger 문서입니다.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5500"
        }
    ],
    "paths": {
...
        "/users/sign-up": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "회원 가입",
                "description": "이메일과 닉네임, 비밀번호를 입력받아 회원 가입을 한다",
                "responses": {
                    "200": {
                        "description": "유저 phone 수정 완료",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "data": {
                                        "phone": "010-0000-0000"
                                    }
                                },
                                "example": {
                                    "data": {
                                        "phone": "010-1111-2222"
                                    }
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "required": true,
                                        "example": "test@test.com"
                                    },
                                    "nickname": {
                                        "type": "string",
                                        "required": true,
                                        "example": "닉네임"
                                    },
                                    "password": {
                                        "type": "string",
                                        "required": true,
                                        "example": "11111111"
                                    }
                                }
                            },
                            "example": {
                                "email": "test@test.com",
                                "nickname": "닉네임",
                                "password": "11111111"
                            }
                        }
                    }
                }
            }
        },
        "/users/log-in": {
            "post": {
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "example": "any"
                                    },
                                    "password": {
                                        "example": "any"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/refresh-token": {
            "post": {
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "refreshToken": {
                                        "example": "any"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    }
}